Removed excalidraw entirely
