+++ apps/frontend/index.html
@@ -80,9 +80,5 @@
       type="module"
       src="./src/main.tsx"
     ></script>
-    <script>
-      const isDev = window.location.hostname === "localhost";
-      window.EXCALIDRAW_ASSET_PATH = isDev ? undefined : "/assets/";
-    </script>
   </body>
 </html>
