+++ libs/src/assets/index.ts
@@ -12,6 +12,6 @@
 
 export { default as FilesharingIcon } from './filesharing.svg';
 export { default as SurveysIcon } from './surveys.svg';
-export { default as WhiteBoardCollaborationIcon } from './whiteboard_collaboration.svg';
+export { default as WhiteBoardIcon } from './whiteboard.svg';
 export { default as ClassManagementIcon } from './classmgmt.svg';
 export { default as BulletinBoardIcon } from './kanban.svg';
