+++ package.json
@@ -33,7 +33,6 @@
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
-    "@excalidraw/excalidraw": "^0.17.6",
     "@hookform/resolvers": "^3.3.4",
     "@nestjs/axios": "^3.0.2",
     "@nestjs/bullmq": "^11.0.2",
